Renamed and reorganized components
