+++ src/components/InputPanel/PanelInputInterval.jsx
